fix: 컨트랙트 코인 양 조회시 DB에 업데이트 하도록 수정

+++ src/main/java/com/baekho/bridgenet/domain/bridge/dto/response/ChainUpdateResponseDTO.java
@@ -3,13 +3,15 @@ package com.baekho.bridgenet.domain.bridge.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 @AllArgsConstructor
 public class ChainUpdateResponseDTO {
     private Long chainId;
     private String chainName;
     private String smartContractAddress;
-    private Long smartContractValue;
+    private BigInteger smartContractValue;
     private String unit;
     private String httpRpc;
     private String wsRpc;
